fit' and unwrap'
`fit’` works like `fit`, but first chops up the source pattern into a
list of patterns.  Then those pieces are put back together according to
the index pattern.

`unwrap’` is in support of this, and flattens out a pattern of patterns
by compressing the inner patterns into the timing of the outer pattern.