Lindenmayer "L-system" patterns
`lindenmayerN iter rule start_string` gives a string that’s the result
of applying the rule string to the start_string `iter` times.  The rule
string should have a format like
`0:1,1:12,2:000`, i.e. rules are comma-separated, and the token on the
left of the colon is replaced with the characters to the right.