Lindenmayer systems
Functions for generating strings of Lindenmayer systems for use in
Tidal.  Especially useful in combination with something that converts a
string to a Pattern (like `step` or `step’`).

The basic function `lindenmayer` takes a rule string of the form and
applies it to an initial sequence, iterating N times.  Rules are
separated by commas, with from/to separated by colons.
For example
`lindenmayer 1 “a:b,b:ab” “ab”` returns `”bab”`

To turn this into a Pattern for Tidal, you can do something like this:
`d1 $ sound $ slow 8 $ step "ifdrums" (take 512 $ lindenmayer 6
"0:1--0---0--1--10,1:11" “0”)`