Sieve data type
Implements Xenakis Sieves as a data type in Tidal.  The basic notation
is
`3@@1` which results in the sequence [1, 4, 7, 10, ...], but the data
type allows for various representations.

Unary and binary logical operators are defined: `not'`, `#&&#` (and),
`#||#` (or), and `#^^#` (xor).

Usage in Tidal is so far limited to two cases:
  1) a finite portion of a sieve can be interpreted as a sequence of
beats in a pattern, somewhat like the Euler rhythms.  For example,
`sieveToPat 2@@0` would result in the pattern `"x ~ x ~"`
  2) the integer indices of a finite portion of a sieve can be treated
as a sequence of notes forming a scale or chord. The function
`scaleSieve` is provided for this.