Fix issue with toScale.
It seemed to be behaving strangely with patterns of overlapping
events, or patterns containing events that spanned multiple cycles.

This version fixes that by mapping over the pattern just once,
rather than doing it once for the note value and once for the octave
and then combining the two.

Also touch up documentation.