Add a Nix Flake for building and developing Tidal
A Nix flake for building and developing Tidal.

Packages are included for:
- tidal
- tidal-link
- tidal-listener
- tidal-parse

A `tidal-ghci` package is also included. This is a small script that
starts an instance of `GHCi` with `Tidal` installed and with the
`BootTidal.hs` file passed as the `-ghci-script`.

Packages can be built with `nix build .#tidal` or ran with `nix run
.#tidal-ghci`.

A `devShell` is included that provides `cabal-install`, `stack` and all
other build inputs for the tidal packages above included under a
temporary shell. This shell can be entered with `nix develop`.